Decompose metadata table projection internals
